+++ server/fontello/api.update.js
@@ -6,6 +6,7 @@
 const _           = require('lodash');
 const validator   = require('is-my-json-valid');
 const read        = require('util').promisify(require('fs').readFile);
+const gunzip      = require('util').promisify(require('zlib').gunzip);
 
 
 const config_schema = require('./font/_lib/config_schema');
@@ -38,7 +39,9 @@ module.exports = function (N, apiPath) {
     let configFile;
 
     // Extract config
-    configFile = await read(configPath, { encoding: 'utf-8' });
+    configFile = await read(configPath);
+
+    configFile = (await gunzip(configFile)).toString();
 
     let config;