+++ server/fontello/font/generate/generate.js
@@ -7,6 +7,8 @@ const _             = require('lodash');
 const validator     = require('is-my-json-valid');
 const config_schema = require('../_lib/config_schema');
 const read          = require('util').promisify(require('fs').readFile);
+const gunzip        = require('util').promisify(require('zlib').gunzip);
+
 
 
 module.exports = function (N, apiPath) {
@@ -23,7 +25,9 @@ module.exports = function (N, apiPath) {
 
     if (!path) throw N.io.BAD_REQUEST;
 
-    let file = await read(path, { encoding: 'utf-8' });
+    let file = await read(path);
+
+    file = (await gunzip(file)).toString();
 
     try {
       env.data.config = JSON.parse(file);